refactor(desktop): place cache boundary with cache owner
